Moves `PolicyHasValidAWSPrincipals` from `sns` to `iam`
